# backend/docs/swagger.json
{
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "AI Agent Demo — Go + Gin 后端，提供 13 个技术演示场景的 API 支持",
    "title": "Interview Demo API",
    "contact": {
      "name": "Developer",
      "email": "dev@example.com"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
...
    "/config": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "返回模拟的集群配置（含名称、副本数、TLS、日志级别），用于演示 LRU 缓存",
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "集群配置",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/schema/config": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "返回基站配置的动态表单结构（含 tabs/cards/leaf 组件、字段属性、校验规则），支持前端动态渲染",
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "获取动态表单 Schema",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/schema/validate": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "根据 JSON Schema 对提交数据进行后端校验，含业务规则（IP 格式、小区 ID、MCC/MNC 等）",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "Schema 数据校验",
        "parameters": [
          {
            "description": "校验请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/demo.ValidationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/upload/chunk": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "上传单个文件分片（multipart/form-data），含 SHA256 哈希校验",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "上传分片",
        "parameters": [
          {
            "type": "string",
            "description": "上传会话 ID",
            "name": "uploadId",
            "in": "formData",
            "required": true
          },
          {
            "type": "integer",
            "description": "分片索引",
            "name": "chunkIndex",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "分片 SHA256 哈希",
            "name": "hash",
            "in": "formData",
            "required": true
          },
          {
            "type": "file",
            "description": "分片文件内容",
            "name": "chunk",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/upload/complete": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "合并所有已上传分片为完整文件，验证文件完整性",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "完成分片上传",
        "parameters": [
          {
            "description": "完成请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/demo.CompleteUploadRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/upload/download/{uploadId}": {
      "get": {
        "description": "根据 uploadId 下载已合并的完整文件",
        "produces": ["application/octet-stream"],
        "tags": ["演示"],
        "summary": "下载已上传文件",
        "parameters": [
          {
            "type": "string",
            "description": "上传会话 ID",
            "name": "uploadId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/upload/init": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "创建新的分片上传会话，返回 uploadId 用于后续分片上传",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "初始化分片上传",
        "parameters": [
          {
            "description": "初始化请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/demo.InitUploadRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/upload/sessions": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "返回当前所有活跃的上传会话列表",
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "列出上传会话",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/upload/status/{uploadId}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "查询指定上传会话的进度和已接收分片列表",
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "查询上传状态",
        "parameters": [
          {
            "type": "string",
            "description": "上传会话 ID",
            "name": "uploadId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/vitals/page-report": {
      "post": {
        "description": "上报前端页面渲染指标（渲染耗时、LCP/INP/CLS），用于页面性能分析",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "上报页面渲染数据",
        "parameters": [
          {
            "description": "页面报告数组",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/demo.PageReport"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
    "/vitals/report": {
      "post": {
        "description": "上报前端 Web Vitals 指标（TTFB/FCP/LCP/CLS/INP），用于实时监控面板",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["演示"],
        "summary": "上报 Web Vitals",
        "parameters": [
          {
            "description": "Vitals 上报数组",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/demo.VitalsReport"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "demo.CompleteUploadRequest": {
      "type": "object",
      "required": ["uploadId"],
      "properties": {
        "uploadId": {
          "type": "string"
        }
      }
    },
    "demo.InitUploadRequest": {
      "type": "object",
      "required": ["chunkSize", "fileSize", "filename", "totalChunks"],
      "properties": {
        "chunkSize": {
          "type": "integer"
        },
        "fileHash": {
          "type": "string"
        },
        "fileSize": {
          "type": "integer"
        },
        "filename": {
          "type": "string"
        },
        "totalChunks": {
          "type": "integer"
        }
      }
    },
    "demo.PageReport": {
      "type": "object",
      "properties": {
        "cls": {
          "type": "number"
        },
        "inp": {
          "type": "number"
        },
        "lcp": {
          "type": "number"
        },
        "pageName": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "referrer": {
          "type": "string"
        },
        "renderDuration": {
          "type": "number"
        }
      }
    },
...
    "demo.ValidationRequest": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "additionalProperties": true
        },
        "schema": {
          "type": "object",
          "additionalProperties": true
        }
      }
    },
    "demo.VitalsReport": {
      "type": "object",
      "properties": {
        "cls": {
          "type": "number"
        },
        "fcp": {
          "type": "number"
        },
        "inp": {
          "type": "number"
        },
        "lcp": {
          "type": "number"
        },
        "metric": {
          "type": "string"
        },
        "rating": {
          "type": "string"
        },
        "ttfb": {
          "type": "number"
        },
        "url": {
          "type": "string"
        },
        "value": {
          "type": "number"
        },
        "version": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "输入格式: \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}